requirements.txt
# Core LLM APIs
openai
anthropic

# Environment management
python-dotenv

# Token counting
tiktoken

# MCP (Model Context Protocol)
mcp>=0.9.0

# Data validation and parsing
pydantic

# HTTP requests
requests

# Vector databases (for advanced tutorials)
chromadb

# Data processing
numpy

# Interactive notebooks
jupyter
ipykernel

# Testing
pytest
pytest-asyncio
pytest-cov

# Code quality
black
flake8
mypy

# Retry logic
tenacity

# JSON schema validation
jsonschema

# Date/time utilities
python-dateutil

# Logging
colorlog

# CLI utilities
rich
click

# Optional: Web UI frameworks
gradio
streamlit

# Optional: Additional LLM frameworks
langchain
langchain-community
langchain-openai
langgraph
# AutoGen / AG2 agent framework (see tutorials/08-autogen)
autogen>=0.10.1

# CrewAI framework
crewai>=0.11.0
crewai-tools>=0.2.0
